+++ src/Form/Utilisateurs/ModificationProfilType.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace App\Form\Utilisateurs;
+
+use App\Entity\Utilisateurs;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Validator\Constraints\IsTrue;
+use Symfony\Component\Validator\Constraints\Length;
+use Symfony\Component\Validator\Constraints\NotBlank;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Validator\Constraints\PasswordStrength;
+
+class ModificationProfilType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('email', EmailType::class, [
+                'disabled' => true
+            ])
+            ->add('prenom')
+            ->add('nom')
+            ->add('password', PasswordType::class, [
+                // instead of being set onto the object directly,
+                // this is read and encoded in the controller
+                'mapped' => false,
+                'required' => false,
+                'attr' => ['autocomplete' => 'new-password'],
+                'constraints' => [
+                    new PasswordStrength([
+                        'minScore' => PasswordStrength::STRENGTH_STRONG,
+                        'message' => 'Le mot de passe est trop facile, veuillez mettre un mot de passe plus compliqué.'
+                    ])
+                ],
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Utilisateurs::class
+        ]);
+    }
+}
